--- resources/jobTemplate.xml
@@ -1,51 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject plugin="workflow-multibranch@2.16">
-    <actions/>
-    <description></description>
-    <properties>
-        <org.jenkinsci.plugins.pipeline.modeldefinition.config.FolderConfig plugin="pipeline-model-definition@1.2.2">
-            <dockerLabel></dockerLabel>
-            <registry plugin="docker-commons@1.9"/>
-        </org.jenkinsci.plugins.pipeline.modeldefinition.config.FolderConfig>
-    </properties>
-    <folderViews class="jenkins.branch.MultiBranchProjectViewHolder" plugin="branch-api@2.0.15">
-        <owner class="org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject" reference="../.."/>
-    </folderViews>
-    <healthMetrics>
-        <com.cloudbees.hudson.plugins.folder.health.WorstChildHealthMetric plugin="cloudbees-folder@6.2.1">
-            <nonRecursive>false</nonRecursive>
-        </com.cloudbees.hudson.plugins.folder.health.WorstChildHealthMetric>
-    </healthMetrics>
-    <icon class="jenkins.branch.MetadataActionFolderIcon" plugin="branch-api@2.0.15">
-        <owner class="org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject" reference="../.."/>
-    </icon>
-    <orphanedItemStrategy class="com.cloudbees.hudson.plugins.folder.computed.DefaultOrphanedItemStrategy" plugin="cloudbees-folder@6.2.1">
-        <pruneDeadBranches>true</pruneDeadBranches>
-        <daysToKeep>-1</daysToKeep>
-        <numToKeep>-1</numToKeep>
-    </orphanedItemStrategy>
-    <triggers/>
-    <disabled>false</disabled>
-    <sources class="jenkins.branch.MultiBranchProject\$BranchSourceList" plugin="branch-api@2.0.15">
-        <data>
-            <jenkins.branch.BranchSource>
-                <source class="jenkins.plugins.git.GitSCMSource" plugin="git@3.6.3">
-                    <id>${guid}</id>
-                    <remote>${gitUri}</remote>
-                    <credentialsId></credentialsId>
-                    <traits>
-                        <jenkins.plugins.git.traits.BranchDiscoveryTrait/>
-                    </traits>
-                </source>
-                <strategy class="jenkins.branch.DefaultBranchPropertyStrategy">
-                    <properties class="empty-list"/>
-                </strategy>
-            </jenkins.branch.BranchSource>
-        </data>
-        <owner class="org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject" reference="../.."/>
-    </sources>
-    <factory class="org.jenkinsci.plugins.workflow.multibranch.WorkflowBranchProjectFactory">
-        <owner class="org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject" reference="../.."/>
-        <scriptPath>Jenkinsfile</scriptPath>
-    </factory>
-</org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject>
